chore: rename cli to accumulate
