+++ cmd/accumulate/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/AccumulateNetwork/accumulate/cmd/cli/cmd"
+	"github.com/AccumulateNetwork/accumulate/cmd/accumulate/cmd"
 )
 
 func main() {
