+++ scripts/matrix/test_case_2.4.sh
@@ -7,13 +7,13 @@
 #
 # set cli command and see if it exists
 #
-export cli=../../cmd/cli/cli
+export cli=../../cmd/accumulate/accumulate
 
 if [ ! -f $cli ]; then
-        echo "cli command not found in ../../cmd/cli, attempting to build"
+        echo "accumulate command not found in ../../cmd/cli, attempting to build"
         ./build_cli.sh
         if [ ! -f $cli ]; then
-           echo "cli command failed to build"
+           echo "accumulate command failed to build"
            exit 1
         fi
 fi
@@ -23,7 +23,7 @@ fi
 ID=`./cli_create_id.sh $1`
 
 if [ $? -ne 0 ]; then
-	echo "cli create id failed"
+	echo "accumulate create id failed"
 	exit 1
 fi
 
@@ -34,7 +34,7 @@ echo $ID
 TxID=`./cli_faucet.sh $ID $1`
 
 if [ $? -ne 0 ]; then
-	echo "cli faucet failed"
+	echo "accumulate faucet failed"
 	exit 1
 fi
 
@@ -45,7 +45,7 @@ sleep 2.5
 bal=`./cli_get_balance.sh $ID $1`
 
 if [ $? -ne 0 ]; then
-	echo "cli get balance failed"
+	echo "accumulate get balance failed"
 	exit 1
 fi
 
@@ -56,7 +56,7 @@ echo $bal
 Key=`./cli_key_generate.sh t24key $1`
 
 if [ $? -ne 0 ]; then
-	echo "cli key generate failed"
+	echo "accumulate key generate failed"
 	exit 1
 fi
 
@@ -67,7 +67,7 @@ echo $key
 ./cli_adi_create_account.sh $ID acc://012345678 t24key $1
 
 if [ $? -eq 0 ]; then
-	echo "cli adi create account passed and it should have failed"
+	echo "accumulate adi create account passed and it should have failed"
 	exit 1
 fi
 exit 0
