+++ scripts/matrix/test_case_2.6.sh
@@ -7,13 +7,13 @@
 #
 # set cli command and see if it exists
 #
-export cli=../../cmd/cli/cli
+export cli=../../cmd/accumulate/accumulate
 
 if [ ! -f $cli ]; then
-        echo "cli command not found in ../../cmd/cli, attempting to build"
+        echo "accumulate command not found in ../../cmd/cli, attempting to build"
         ./build_cli.sh
         if [ ! -f $cli ]; then
-           echo "cli command failed to build"
+           echo "accumulate command failed to build"
            exit 1
         fi
 fi
@@ -23,7 +23,7 @@ fi
 ID=`./cli_create_id.sh $1`
 
 if [ $? -ne 0 ]; then
-	echo "cli create id failed"
+	echo "accumulate create id failed"
 	exit 1
 fi
 echo $ID
@@ -33,7 +33,7 @@ echo $ID
 TxID=`./cli_faucet.sh $ID $1`
 
 if [ $? -ne 0 ]; then
-	echo "cli faucet failed"
+	echo "accumulate faucet failed"
 	exit 1
 fi
 
@@ -44,7 +44,7 @@ sleep 2.5
 bal=`./cli_get_balance.sh $ID $1`
 
 if [ $? -ne 0 ]; then
-	echo "cli get balance failed"
+	echo "accumulate get balance failed"
 	exit 1
 fi
 echo $bal
@@ -53,12 +53,12 @@ echo $bal
 
 Key=`./cli_key_generate.sh t26key $1`
 if [ $? -ne 0 ]; then
-	echo "cli key generate failed"
+	echo "accumulate key generate failed"
 	exit 1
 fi
 Key=`./cli_key_generate.sh t262key $1`
 if [ $? -ne 0 ]; then
-	echo "cli key generate failed"
+	echo "accumulate key generate failed"
 	exit 1
 fi
 
@@ -68,7 +68,7 @@ echo $key
 
 ./cli_adi_create_account.sh $ID acc://t26acct t26key $1
 if [ $? -ne 0 ]; then
-	echo "cli adi create account failed"
+	echo "accumulate adi create account failed"
 	exit 1
 fi
 
@@ -78,7 +78,7 @@ sleep 2.5
 
 ./cli_adi_create_account.sh acc://t26acct acc://t262acct t262key $1
 if [ $? -ne 0 ]; then
-	echo "cli adi create account failed"
+	echo "accumulate adi create account failed"
 	exit 1
 fi
