+++ scripts/matrix/test_case_3.1.sh
@@ -7,13 +7,13 @@
 #
 # set cli command and see if it exists
 #
-export cli=../../cmd/cli/cli
+export cli=../../cmd/accumulate/accumulate
 
 if [ ! -f $cli ]; then
-        echo "cli command not found in ../../cmd/cli, attempting to build"
+        echo "accumulate command not found in ../../cmd/cli, attempting to build"
         ./build_cli.sh
         if [ ! -f $cli ]; then
-           echo "cli command failed to build"
+           echo "accumulate command failed to build"
            exit 1
         fi
 fi
@@ -22,7 +22,7 @@ fi
 #
 ID=`./cli_create_id.sh $1`
 if [ $? -ne 0 ]; then
-	echo "cli create id failed"
+	echo "accumulate create id failed"
 	exit 1
 fi
 
@@ -32,7 +32,7 @@ echo $ID
 
 TxID=`./cli_faucet.sh $ID $1`
 if [ $? -ne 0 ]; then
-	echo "cli faucet failed"
+	echo "accumulate faucet failed"
 	exit 1
 fi
 
@@ -41,7 +41,7 @@ fi
 sleep 2.5
 bal=`./cli_get_balance.sh $ID $1`
 if [ $? -ne 0 ]; then
-	echo "cli get balance failed"
+	echo "accumulate get balance failed"
 	exit 1
 fi
 
@@ -51,7 +51,7 @@ echo $bal
 
 Key=`./cli_key_generate.sh t31key $1`
 if [ $? -ne 0 ]; then
-	echo "cli key generate failed"
+	echo "accumulate key generate failed"
 	exit 1
 fi
 
@@ -61,14 +61,14 @@ echo $key
 
 ./cli_adi_create_account.sh $ID acc://t31acct t31key $1
 if [ $? -ne 0 ]; then
-	echo "cli adi create account failed"
+	echo "accumulate adi create account failed"
 	exit 1
 fi
 
 sleep 2.5
 $cli account create token acc://t31acct t31key acc://t31acct/myacmeacct acc://ACME acc://t31acct/book0 -s http://$1/v1
 if [ $? -ne 0 ]; then
-	echo "cli account create failed"
+	echo "accumulate account create failed"
 	exit 1
 fi
 exit 0
