+++ scripts/matrix/test_case_3.2.sh
@@ -7,13 +7,13 @@
 #
 # set cli command and see if it exists
 #
-export cli=../../cmd/cli/cli
+export cli=../../cmd/accumulate/accumulate
 
 if [ ! -f $cli ]; then
-        echo "cli command not found in ../../cmd/cli, attempting to build"
+        echo "accumulate command not found in ../../cmd/cli, attempting to build"
         ./build_cli.sh
         if [ ! -f $cli ]; then
-           echo "cli command failed to build"
+           echo "accumulate command failed to build"
            exit 1
         fi
 fi
@@ -23,7 +23,7 @@ fi
 ID=`./cli_create_id.sh $1`
 
 if [ $? -ne 0 ]; then
-	echo "cli create id failed"
+	echo "accumulate create id failed"
 	exit 1
 fi
 echo $ID
@@ -33,7 +33,7 @@ echo $ID
 TxID=`./cli_faucet.sh $ID $1`
 
 if [ $? -ne 0 ]; then
-	echo "cli faucet failed"
+	echo "accumulate faucet failed"
 	exit 1
 fi
 # get our balance
@@ -41,7 +41,7 @@ fi
 sleep 2.5
 bal=`./cli_get_balance.sh $ID $1`
 if [ $? -ne 0 ]; then
-	echo "cli get balance failed"
+	echo "accumulate get balance failed"
 	exit 1
 fi
 
@@ -51,7 +51,7 @@ echo $bal
 
 Key=`./cli_key_generate.sh t32key $1`
 if [ $? -ne 0 ]; then
-	echo "cli key generate failed"
+	echo "accumulate key generate failed"
 	exit 1
 fi
 
@@ -62,7 +62,7 @@ echo $key
 sleep 2.5
 $cli account create token acc://t32acct t32key acc://t32acct/myacmeacct acc://ACME acc://t32acct/book0 -s http://$1/v1
 if [ $? -eq 0 ]; then
-	echo "cli account create passed and should have failed"
+	echo "accumulate account create passed and should have failed"
 	exit 1
 fi
 exit 0
