+++ scripts/matrix/test_case_4.3.sh
@@ -7,7 +7,7 @@
 #
 # set cli command and see if it exists
 #
-export cli=../../cmd/cli/cli
+export cli=../../cmd/accumulate/accumulate
 
 if [ -z $1 ]; then
 	echo "must supply host:port"
@@ -15,10 +15,10 @@ if [ -z $1 ]; then
 fi
 
 if [ ! -f $cli ]; then
-	echo "cli command not found in ../../cmd/cli, attempting to build"
+	echo "accumulate command not found in ../../cmd/accumulate, attempting to build"
         ./build_cli.sh
         if [ ! -f $cli ]; then
-           echo "cli command failed to build"
+           echo "accumulate command failed to build"
            exit 1
         fi
 fi
@@ -28,7 +28,7 @@ fi
 
 ID=`./cli_create_id.sh $1`
 if [ $? -ne 0 ]; then
-	echo "cli create id failed"
+	echo "accumulate create id failed"
 	exit 1
 fi
 
@@ -40,7 +40,7 @@ sleep .5
 
 TxID=`./cli_faucet.sh $ID $1`
 if [ $? -ne 0 ]; then
-	echo "cli faucet failed"
+	echo "accumulate faucet failed"
 	exit 1
 fi
 
@@ -48,17 +48,17 @@ fi
 
 Key=`./cli_key_generate.sh t43key $1`
 if [ $? -ne 0 ]; then
-	echo "cli key generate failed"
+	echo "accumulate key generate failed"
 	exit 1
 fi
 Key2=`./cli_key_generate.sh t43key2 $1`
 if [ $? -ne 0 ]; then
-	echo "cli key generate failed"
+	echo "accumulate key generate failed"
 	exit 1
 fi
 Key3=`./cli_key_generate.sh t43key3 $1`
 if [ $? -ne 0 ]; then
-	echo "cli key generate failed"
+	echo "accumulate key generate failed"
 	exit 1
 fi
 
@@ -67,14 +67,14 @@ fi
 sleep 2
 ./cli_adi_create_account.sh $ID acc://t43acct t43key $1
 if [ $? -ne 0 ]; then
-	echo "cli adi create account failed"
+	echo "accumulate adi create account failed"
 	exit 1
 fi
 
 sleep 2
 $cli page create acc://t43acct t43key acc://t43acct/keypage43 t43key2 -s http://$1/v1
 if [ $? -ne 0 ]; then
-	echo "cli page create failed"
+	echo "accumulate page create failed"
 	exit 1
 fi
 
@@ -82,7 +82,7 @@ sleep 2
 $cli page key add acc://t43acct/keypage43 t43key t43key3 -s http://$1/v1
 
 if [ $? -ne 0 ]; then
-	echo "cli page key add failed"
+	echo "accumulate page key add failed"
 	exit 1
 fi
 exit 0
