+++ scripts/matrix/test_case_6.2.sh
@@ -7,13 +7,13 @@
 #
 # set cli command and see if it exists
 #
-export cli=../../cmd/cli/cli
+export cli=../../cmd/accumulate/accumulate
 
 if [ ! -f $cli ]; then
-        echo "cli command not found in ../../cmd/cli, attempting to build"
+        echo "accumulate command not found in ../../cmd/cli, attempting to build"
         ./build_cli.sh
         if [ ! -f $cli ]; then
-           echo "cli command failed to build"
+           echo "accumulate command failed to build"
            exit 1
         fi
 fi
@@ -58,7 +58,7 @@ fi
 
 ./cli_xfer_tokens.sh $id1 $id2 $3 $4
 if [ $? -ne 0 ]; then
-	echo "cli xfer tokens failed"
+	echo "accumulate xfer tokens failed"
 	exit 1
 fi
 exit 0
